Supervise server process on daily basis
Starting server process only when certain preconditions were met seems to be
a faulty idea. First of all, if for some reasons someone killed this process,
cmus has had to be restarted in order to restore cmusfm functionality. But
the real reason why this idea was bad, is fact, that the behavior of status
display program invocation might vary if resume option is enabled in cmus.

Fixes #3