Use POSIX message queue for our IPC
Drop socket-based communication if favor of the POSIX message queue. The
main advantage of using this approach is code simplification and reduction
of the communication overhead - the network stack. Note, that we are using
a non-portable implementation which allows multiplexing, which means that
this code will not work in the Cygwin environment...