Revert of "Use POSIX message queue..."
This reverts commit 5ac95e4.

It has turned out, that the POSIX message queue is a little bit too young to
serve our purpose. The feasible implementation (from our point of view) is
available on the Linux box only. FreeBSD implementation is in fact also based
on file descriptors, however the message queue handler is a structure which
holds file descriptor internally... not to mention, that prior to usage, the
message queue file system has to be mounted.

In the simplicity vs portability contest, the portability wins (at least in
our case).